Move r tasks to new module
